pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "arabic-reshaper>=3.0.0",
    "faiss-cpu>=1.11.0",
    "google-generativeai>=0.8.5",
    "langchain-community>=0.3.23",
    "langchain>=0.3.25",
    "pydub>=0.25.1",
    "pymupdf>=1.25.5",
    "python-bidi>=0.6.6",
    "reportlab>=4.4.0",
    "streamlit>=1.45.0",
    "tensorflow-hub>=0.16.1",
    "whisper>=1.1.10",
    "tensorflow>=2.19.0",
    "numpy>=2.1.3",
]
